fix: address clippy errors and review comments
- Fix redundant else block in bench_reproduce (clippy::redundant_else)
- Rename instance_drifted → any_step_drifted in reproduce.rs to avoid
  name-similarity lint with instances_drifted
- Simplify sampling_differs in compare.rs to use PartialEq (a != b),
  now also covers the extra field
- Simplify reproduce.rs step comparison to use ss != rs (PartialEq)
- Refactor load_trajectory_for_instance into trajectory::mod as a
  shared pub fn used by both compare.rs and reproduce.rs
- Remove unused imports (BTreeMap, ExitReason, FilterSpec) from
  sampling_params tests; add allow(expect_used) to test allow list
- Fix field-assignment-outside-initializer by using struct literal
  with ..Default::default() in write_traj_with_temperature
- Remove redundant .clone() calls in CompareArgs construction
- Add #[allow(dead_code)] to support/mod.rs binary path helpers
- Update docs/spec-trajectory.md to remove "(future work)" note
  since --strict-sampling is now fully implemented
